Redesigned the emitter
The emitter should now be easier to comprehend. The trick is that parent
expressions adds a :db/id key to it's child expressions. As before, the
emitter returns transaction data, but now in the form of maps.

Another substantial change is that I now derive 'child' (again) instead
of explicitly recording it in the database as :ast/child